+++ packages/builder/src/utils/eslint-utils.spec.ts
@@ -1,18 +1,9 @@
-// Force module scoping
-export default {};
-
 jest.mock('eslint', () => ({
   ESLint: jest.fn(),
 }));
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { ESLint } = require('eslint');
-(<jest.SpyInstance>ESLint).mockImplementation(() => ({
-  lintFiles: (args: string[]) => args,
-}));
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { lint } = require('./eslint-utils');
+import { ESLint } from 'eslint';
+import { resolveAndInstantiateESLint } from './eslint-utils';
 
 describe('eslint-utils', () => {
   beforeEach(() => {
@@ -20,13 +11,12 @@ describe('eslint-utils', () => {
   });
 
   it('should create the ESLint instance with the proper parameters', async () => {
-    await lint('/root', './.eslintrc.json', {
+    await resolveAndInstantiateESLint('./.eslintrc.json', {
       fix: true,
       cache: true,
       cacheLocation: '/root/cache',
       cacheStrategy: 'content',
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-    }).catch(() => {});
+    } as any);
 
     expect(ESLint).toHaveBeenCalledWith({
       overrideConfigFile: './.eslintrc.json',
@@ -42,13 +32,12 @@ describe('eslint-utils', () => {
   });
 
   it('should create the ESLint instance with the proper parameters', async () => {
-    await lint('/root', undefined, {
+    await resolveAndInstantiateESLint(undefined, {
       fix: true,
       cache: true,
       cacheLocation: '/root/cache',
       cacheStrategy: 'content',
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-    }).catch(() => {});
+    } as any);
 
     expect(ESLint).toHaveBeenCalledWith({
       overrideConfigFile: undefined,
@@ -65,13 +54,12 @@ describe('eslint-utils', () => {
 
   describe('noEslintrc', () => {
     it('should create the ESLint instance with "useEslintrc" set to false', async () => {
-      await lint('/root', undefined, {
+      await resolveAndInstantiateESLint(undefined, {
         fix: true,
         cache: true,
         cacheLocation: '/root/cache',
         noEslintrc: true,
-        // eslint-disable-next-line @typescript-eslint/no-empty-function
-      }).catch(() => {});
+      } as any);
 
       expect(ESLint).toHaveBeenCalledWith({
         overrideConfigFile: undefined,
@@ -89,14 +77,13 @@ describe('eslint-utils', () => {
   describe('rulesdir', () => {
     it('should create the ESLint instance with "rulePaths" set to the given value for rulesdir', async () => {
       const extraRuleDirectories = ['./some-rules', '../some-more-rules'];
-      await lint('/root', undefined, {
+      await resolveAndInstantiateESLint(undefined, {
         fix: true,
         cache: true,
         cacheLocation: '/root/cache',
         cacheStrategy: 'content',
         rulesdir: extraRuleDirectories,
-        // eslint-disable-next-line @typescript-eslint/no-empty-function
-      }).catch(() => {});
+      } as any);
 
       expect(ESLint).toHaveBeenCalledWith({
         fix: true,
@@ -113,14 +100,13 @@ describe('eslint-utils', () => {
 
   describe('resolvePluginsRelativeTo', () => {
     it('should create the ESLint instance with "resolvePluginsRelativeTo" set to the given value for resolvePluginsRelativeTo', async () => {
-      await lint('/root', undefined, {
+      await resolveAndInstantiateESLint(undefined, {
         fix: true,
         cache: true,
         cacheLocation: '/root/cache',
         cacheStrategy: 'content',
         resolvePluginsRelativeTo: './some-path',
-        // eslint-disable-next-line @typescript-eslint/no-empty-function
-      }).catch(() => {});
+      } as any);
 
       expect(ESLint).toHaveBeenCalledWith({
         fix: true,
@@ -133,6 +119,78 @@ describe('eslint-utils', () => {
         rulePaths: [],
         resolvePluginsRelativeTo: './some-path',
       });
+    });
+  });
+
+  describe('reportUnusedDisableDirectives', () => {
+    it('should create the ESLint instance with "reportUnusedDisableDirectives" set to the given value for reportUnusedDisableDirectives', async () => {
+      await resolveAndInstantiateESLint(undefined, {
+        fix: true,
+        cache: true,
+        cacheLocation: '/root/cache',
+        cacheStrategy: 'content',
+        reportUnusedDisableDirectives: 'warn',
+      } as any);
+
+      expect(ESLint).toHaveBeenCalledWith({
+        fix: true,
+        cache: true,
+        cacheLocation: '/root/cache',
+        cacheStrategy: 'content',
+        ignorePath: undefined,
+        useEslintrc: true,
+        errorOnUnmatchedPattern: false,
+        rulePaths: [],
+        reportUnusedDisableDirectives: 'warn',
+      });
+    });
+  });
+
+  describe('ESLint Flat Config', () => {
+    it('should throw if a non eslint.config.js file is used with ESLint Flat Config', async () => {
+      await expect(
+        resolveAndInstantiateESLint('./.eslintrc.json', {} as any, true),
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        `"When using the new Flat Config with ESLint, all configs must be named eslint.config.js and .eslintrc files may not be used. See https://eslint.org/docs/latest/use/configure/configuration-files-new"`,
+      );
+    });
+
+    it('should throw if invalid options are used with ESLint Flat Config', async () => {
+      await expect(
+        resolveAndInstantiateESLint(
+          undefined,
+          {
+            useEslintrc: false,
+          } as any,
+          true,
+        ),
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        `"For Flat Config, the \`useEslintrc\` option is not applicable. See https://eslint.org/docs/latest/use/configure/configuration-files-new"`,
+      );
+
+      await expect(
+        resolveAndInstantiateESLint(
+          undefined,
+          {
+            resolvePluginsRelativeTo: './some-path',
+          } as any,
+          true,
+        ),
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        `"For Flat Config, ESLint removed \`resolvePluginsRelativeTo\` and so it is not supported as an option. See https://eslint.org/docs/latest/use/configure/configuration-files-new"`,
+      );
+
+      await expect(
+        resolveAndInstantiateESLint(
+          undefined,
+          {
+            ignorePath: './some-path',
+          } as any,
+          true,
+        ),
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        `"For Flat Config, ESLint removed \`ignorePath\` and so it is not supported as an option. See https://eslint.org/docs/latest/use/configure/configuration-files-new"`,
+      );
     });
   });
 });