+++ packages/eslint-plugin-template/src/rules/click-events-have-key-events.ts
@@ -16,7 +16,7 @@ export default createESLintRule<Options, MessageIds>({
     type: 'suggestion',
     docs: {
       description:
-        'Ensures that the click event is accompanied with at least one key event keyup, keydown or keypress.',
+        '[Accessibility] Ensures that the click event is accompanied with at least one key event keyup, keydown or keypress.',
       recommended: false,
     },
     schema: [],