+++ packages/schematics/src/add-eslint-to-project/index.ts
@@ -1,9 +1,9 @@
-import type { Rule, Tree } from '@angular-devkit/schematics';
-import { chain } from '@angular-devkit/schematics';
+import type { Tree } from '@nx/devkit';
+import { convertNxGenerator } from '@nx/devkit';
 import {
-  addESLintTargetToProject,
-  createESLintConfigForProject,
-  determineTargetProjectName,
+  addESLintTargetToProject__NX,
+  createESLintConfigForProject__NX,
+  determineTargetProjectName__NX,
 } from '../utils';
 
 interface Schema {
@@ -11,27 +11,25 @@ interface Schema {
   setParserOptionsProject?: boolean;
 }
 
-export default function addESLintToProject(schema: Schema): Rule {
-  return (tree: Tree) => {
-    const projectName = determineTargetProjectName(tree, schema.project);
-    if (!projectName) {
-      throw new Error(
-        '\n' +
-          `
+export default convertNxGenerator(async (tree: Tree, options: Schema) => {
+  const projectName = determineTargetProjectName__NX(tree, options.project);
+  if (!projectName) {
+    throw new Error(
+      '\n' +
+        `
 Error: You must specify a project to add ESLint to because you have multiple projects in your angular.json
 
 E.g. npx ng g @angular-eslint/schematics:add-eslint-to-project {{YOUR_PROJECT_NAME_GOES_HERE}}
-        `.trim(),
-      );
-    }
-    return chain([
-      // Set the lint builder and config in angular.json
-      addESLintTargetToProject(projectName, 'lint'),
-      // Create the ESLint config file for the project
-      createESLintConfigForProject(
-        projectName,
-        schema.setParserOptionsProject ?? false,
-      ),
-    ]);
-  };
-}
+      `.trim(),
+    );
+  }
+
+  // Update the lint builder and config in angular.json
+  addESLintTargetToProject__NX(tree, projectName, 'lint');
+
+  createESLintConfigForProject__NX(
+    tree,
+    projectName,
+    options.setParserOptionsProject ?? false,
+  );
+});