+++ packages/schematics/tests/add-eslint-to-project/index.test.ts
@@ -81,9 +81,11 @@ describe('add-eslint-to-project', () => {
       project: rootProjectName,
     };
 
-    await schematicRunner
-      .runSchematicAsync('add-eslint-to-project', options, appTree)
-      .toPromise();
+    await schematicRunner.runSchematic(
+      'add-eslint-to-project',
+      options,
+      appTree,
+    );
 
     expect(
       readJsonInTree(appTree, 'angular.json').projects[rootProjectName]
@@ -138,6 +140,7 @@ describe('add-eslint-to-project', () => {
           Object {
             "extends": Array [
               "plugin:@angular-eslint/template/recommended",
+              "plugin:@angular-eslint/template/accessibility",
             ],
             "files": Array [
               "*.html",
@@ -155,9 +158,11 @@ describe('add-eslint-to-project', () => {
       project: legacyProjectName,
     };
 
-    await schematicRunner
-      .runSchematicAsync('add-eslint-to-project', options, appTree)
-      .toPromise();
+    await schematicRunner.runSchematic(
+      'add-eslint-to-project',
+      options,
+      appTree,
+    );
 
     const projectConfig = readJsonInTree(appTree, 'angular.json').projects[
       legacyProjectName
@@ -222,9 +227,11 @@ describe('add-eslint-to-project', () => {
       project: otherProjectName,
     };
 
-    await schematicRunner
-      .runSchematicAsync('add-eslint-to-project', options, appTree)
-      .toPromise();
+    await schematicRunner.runSchematic(
+      'add-eslint-to-project',
+      options,
+      appTree,
+    );
 
     const projectConfig = readJsonInTree(appTree, 'angular.json').projects[
       otherProjectName
@@ -291,9 +298,11 @@ describe('add-eslint-to-project', () => {
         setParserOptionsProject: true,
       };
 
-      await schematicRunner
-        .runSchematicAsync('add-eslint-to-project', options, appTree)
-        .toPromise();
+      await schematicRunner.runSchematic(
+        'add-eslint-to-project',
+        options,
+        appTree,
+      );
 
       const projectConfig = readJsonInTree(appTree, 'angular.json').projects[
         legacyProjectName
@@ -365,9 +374,11 @@ describe('add-eslint-to-project', () => {
         setParserOptionsProject: true,
       };
 
-      await schematicRunner
-        .runSchematicAsync('add-eslint-to-project', options, appTree)
-        .toPromise();
+      await schematicRunner.runSchematic(
+        'add-eslint-to-project',
+        options,
+        appTree,
+      );
 
       const projectConfig = readJsonInTree(appTree, 'angular.json').projects[
         otherProjectName
@@ -499,9 +510,11 @@ describe('add-eslint-to-project', () => {
         project: rootProjectName,
       };
 
-      await schematicRunner
-        .runSchematicAsync('add-eslint-to-project', options, tree2)
-        .toPromise();
+      await schematicRunner.runSchematic(
+        'add-eslint-to-project',
+        options,
+        tree2,
+      );
 
       expect(
         readJsonInTree(tree2, 'angular.json').projects[rootProjectName]
@@ -556,6 +569,7 @@ describe('add-eslint-to-project', () => {
             Object {
               "extends": Array [
                 "plugin:@angular-eslint/template/recommended",
+                "plugin:@angular-eslint/template/accessibility",
               ],
               "files": Array [
                 "*.html",